DataViews Extensibility: Allow unregistering the view post revisions action
